# apps/web/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HoldDecision": {
        "properties": {
          "decision": {
            "enum": [
              "approve",
              "decline"
            ],
            "title": "Decision",
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "title": "HoldDecision",
        "type": "object"
      },
...
      "HoldRequestOut": {
        "properties": {
          "contact": {
            "description": "Channels the buyer currently allows",
            "items": {
              "enum": [
                "email",
                "sms",
                "voice"
              ],
              "type": "string"
            },
            "title": "Contact",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decided_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lot_id": {
            "format": "uuid",
            "title": "Lot Id",
            "type": "string"
          },
          "lot_number": {
            "title": "Lot Number",
            "type": "string"
          },
          "lot_status": {
            "$ref": "#/components/schemas/LotStatus"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "declined",
              "withdrawn"
            ],
            "title": "Status",
            "type": "string"
          },
          "subdivision_name": {
            "title": "Subdivision Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "lot_id",
          "lot_number",
          "lot_status",
          "subdivision_name",
          "name",
          "email",
          "phone",
          "message",
          "status",
          "contact",
          "created_at",
          "decided_at"
        ],
        "title": "HoldRequestOut",
        "type": "object"
      },
...
      "InquiryOut": {
        "properties": {
          "contact": {
            "description": "Channels the buyer currently allows",
            "items": {
              "enum": [
                "email",
                "sms",
                "voice"
              ],
              "type": "string"
            },
            "title": "Contact",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lot_id": {
            "format": "uuid",
            "title": "Lot Id",
            "type": "string"
          },
          "lot_number": {
            "title": "Lot Number",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "signed_in": {
            "title": "Signed In",
            "type": "boolean"
          },
          "subdivision_name": {
            "title": "Subdivision Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "lot_id",
          "lot_number",
          "subdivision_name",
          "name",
          "email",
          "phone",
          "message",
          "signed_in",
          "contact",
          "created_at"
        ],
        "title": "InquiryOut",
        "type": "object"
      },
...
      "LotStatus": {
        "enum": [
          "available",
          "on_hold",
          "sold"
        ],
        "title": "LotStatus",
        "type": "string"
      },
...
      "Tenant": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "owner",
              "staff"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "Tenant",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cornerpin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tenants/{tenant_id}": {
      "get": {
        "operationId": "tenant_v1_tenants__tenant_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tenant"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not a member of this tenant"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tenant",
        "tags": [
          "accounts"
        ]
      }
    },
...
    "/v1/tenants/{tenant_id}/hold-requests": {
      "get": {
        "description": "Pending first, then newest first.",
        "operationId": "list_hold_requests_v1_tenants__tenant_id__hold_requests_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HoldRequestOut"
                  },
                  "title": "Response List Hold Requests V1 Tenants  Tenant Id  Hold Requests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found, or not your tenant"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Hold Requests",
        "tags": [
          "leads"
        ]
      }
    },
    "/v1/tenants/{tenant_id}/hold-requests/{hold_id}/decision": {
      "post": {
        "operationId": "decide_hold_request_v1_tenants__tenant_id__hold_requests__hold_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "hold_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Hold Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HoldDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HoldRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found, or not your tenant"
          },
          "409": {
            "description": "Already decided, or the lot is sold"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Hold Request",
        "tags": [
          "leads"
        ]
      }
    },
    "/v1/tenants/{tenant_id}/inquiries": {
      "get": {
        "description": "Newest first.",
        "operationId": "list_inquiries_v1_tenants__tenant_id__inquiries_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "cp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InquiryOut"
                  },
                  "title": "Response List Inquiries V1 Tenants  Tenant Id  Inquiries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found, or not your tenant"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Inquiries",
        "tags": [
          "leads"
        ]
      }
    },
...
  }
}